+++ ipypublish/sphinx/notebook/directives.py
@@ -37,6 +37,28 @@ class NbInfo(NbAdmonition):
     _class = 'note'
 
 
+class NBInputToggle(rst.Directive):
+    """ a toggle button """
+
+    required_arguments = 0
+    optional_arguments = 0
+    option_spec = {}
+    has_content = True  # button text
+
+    def run(self):
+        """This is called by the reST parser."""
+        node = nodes.container()
+        node['classes'].append('nbinput-toggle-all')
+        if self.content:
+            self.state.nested_parse(self.content, self.content_offset, node)
+        else:
+            text = self.arguments[0] if self.arguments and self.arguments[0] else 'Toggle Input Cells'
+            paragraph = nodes.paragraph(text=text)
+            node += paragraph
+
+        return [node]
+
+
 class NbInput(rst.Directive):
     """A notebook input cell with prompt and code area."""
 
@@ -160,7 +182,7 @@ def _create_nbcell_nodes(directive):
         else:
             outer_node += codearea_node
 
-    if isinstance(directive, NbInput) and (config.ipysphinx_code_toggle or 'add-toggle' in directive.options):
+    if isinstance(directive, NbInput) and (config.ipysphinx_input_toggle or 'add-toggle' in directive.options):
         outer_node += sphinx.addnodes.only(
             '', docutils.nodes.container(classes=['toggle-nbinput', 'empty']), expr='html')