+++ ipypublish/sphinx/tests/sourcedirs/notebook_cell_decor/conf.py
@@ -4,5 +4,4 @@ extensions = [
 exclude_patterns = ['_build']
 master_doc = 'contents'
 ipysphinx_show_prompts = True
-ipysphinx_code_toggle = True
-ipysphinx_code_hide = True
+ipysphinx_input_toggle = True